requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
python-dotenv==1.1.1
beautifulsoup4==4.13.4
playwright==1.54.0
langchain==0.3.27
langchain-openai==0.3.32
langchain-chroma==1.1.0
chromadb==1.5.5
